+++ app/src/main/java/com/data/remote/dto/CharacterInfoDto.kt
@@ -1,6 +1,5 @@
 package com.data.remote.dto
 
-import com.domain.model.characterItem
 import com.google.gson.annotations.SerializedName
 
 data class CharacterInfoDto (
