--- app/src/main/java/com/example/dao/CharacterDao.kt
@@ -1,16 +1,0 @@
-package com.example.dao
-
-import androidx.lifecycle.LiveData
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.dto.CharacterDto
-
-interface CharacterDao {
-//    @Insert(onConflict = OnConflictStrategy.REPLACE)
-//    suspend fun insert(characterDto: com.data.remote.dto.CharacterDto)
-//
-//    @Query("SELECT * FROM character_table")
-//    fun getCharacterById(): LiveData<CharacterDto>
-
-}
