+++ app/src/main/java/com/example/myapplication/viewmodel/CharacterInfoViewModel.kt
@@ -1,20 +1,13 @@
-package com.example.myapplication
+package com.example.myapplication.viewmodel
 
 import android.os.Build
 import android.util.Log
 import androidx.annotation.RequiresExtension
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.common.Resource
-import com.data.remote.NeopleApiService
-import com.data.remote.dto.CharacterResponse
-import com.data.remote.dto.Item
-import com.domain.model.characterItem
-import com.domain.use_case.GetCharacterEquipmentUseCase
+import com.domain.model.characterDto
 import com.domain.use_case.GetCharacterInfoUseCase
-import com.domain.use_case.GetCharacterSettingUseCase
 import com.presentation.CharacterListState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -25,36 +18,53 @@ import kotlinx.coroutines.flow.onEach
 import javax.inject.Inject
 
 @HiltViewModel
-class CharacterEquipmentViewModel @Inject constructor(
-    private val getCharacterEquipmentUseCase: GetCharacterEquipmentUseCase
+class CharacterInfoViewModel @Inject constructor(
+    private val getCharacterInfoUseCase: GetCharacterInfoUseCase,
 ) : ViewModel() {
+
+    private val _character = MutableStateFlow<characterDto?>(null)
+    val character: StateFlow<characterDto?> = _character
 
     private val _characterId = MutableStateFlow<List<String?>>(emptyList())
     val characterId: StateFlow<List<String?>> = _characterId.asStateFlow()
 
-    private val _equipment = MutableStateFlow<List<Item>>(emptyList())
-    val equipment: StateFlow<List<Item>> = _equipment
+    private val _characterName =MutableStateFlow<List<String?>>(emptyList())
+    val characterName: StateFlow<List<String?>> = _characterName
+
+    private val _jobName = MutableStateFlow("")
+    val jobName : StateFlow<String> = _jobName
+
+    private val _serverId = MutableStateFlow("")
+    val serverId: StateFlow<String> = _serverId
+
+    private val _level = MutableStateFlow<Int>(0)
+    val level: StateFlow<Int> = _level
 
     @RequiresExtension(extension = Build.VERSION_CODES.S, version = 7)
-    fun getCharacterEquipment(serverId: String, characterId: String) {
-        getCharacterEquipmentUseCase(serverId, characterId).onEach { resource ->
+    fun getCharacterInfo(serverId: String, characterNameItem: String) {
+        getCharacterInfoUseCase(serverId, characterNameItem).onEach { resource ->
             when (resource) {
                 is Resource.Success -> {
                     val characterResponse = resource.data
-                    if (characterResponse != null) {
-                        _equipment.value = characterResponse.equipment
-                    }
+                    val servers = characterResponse?.charactItem
+                    val characterIds = servers?.map { it.characterId }
+                    Log.d("viewmodel", characterIds.toString())
+                    val characterName = servers?.map { it.characterName}
+                    if(characterIds != null) {_characterId.value = characterIds}
+                    if(characterName != null){ _characterName.value = characterName }
                 }
                 is Resource.Error -> {
                     CharacterListState(
                         error = resource.message ?: "An unexpected error occurred"
                     ).toString()
                 }
+
                 is Resource.Loading -> {
                     CharacterListState(
                         isLoading = resource.message ?: "data Loading..."
                     ).toString()
                 }
+                else -> {}
             }
         }.launchIn(viewModelScope)
     }
